Split cli.py into multiple context-related files for maintainability
[no ci integration]
