+++ raiden/ui/cli.py
@@ -5,86 +5,31 @@ gevent.monkey.patch_all()
 if True:
     import json
     import os
-    import signal
     import sys
-    import textwrap
     import traceback
-    from binascii import hexlify
-    from copy import deepcopy
-    from datetime import datetime
-    from itertools import count
     from pathlib import Path
     from subprocess import DEVNULL
-    from tempfile import NamedTemporaryFile, mktemp
-    from typing import Any, Dict
-    from urllib.parse import urljoin, urlparse
+    from tempfile import mktemp
+    from urllib.parse import urljoin
 
     import click
-    import filelock
     import gevent
     import gevent.monkey
-    import requests
     import structlog
     from eth_utils import (
-        denoms,
         to_canonical_address,
         to_checksum_address,
-        to_int,
-        to_normalized_address,
     )
-    from gevent.event import AsyncResult
     from mirakuru import ProcessExitedWithError
-    from requests.exceptions import (
-        ConnectionError as RequestsConnectionError,
-        ConnectTimeout,
-        RequestException,
-    )
-    from web3 import HTTPProvider, Web3
-
-    from raiden.constants import (
-        DISCOVERY_TX_GAS_LIMIT,
-        SQLITE_MIN_REQUIRED_VERSION,
-    )
-    from raiden.accounts import AccountManager
     from raiden.api.rest import APIServer, RestAPI
-    from raiden.exceptions import (
-        AddressWithoutCode,
-        AddressWrongContract,
-        APIServerPortInUseError,
-        ContractVersionMismatch,
-        EthNodeCommunicationError,
-        RaidenError,
-        RaidenServicePortInUseError,
-        ReplacementTransactionUnderpriced,
-        TransactionAlreadyPending,
-    )
+
     from raiden.log_config import configure_logging
-    from raiden.network.blockchain_service import BlockChainService
-    from raiden.network.discovery import ContractDiscovery
-    from raiden.network.rpc.client import JSONRPCClient
     from raiden.network.sockfactory import SocketFactory
-    from raiden.network.throttle import TokenBucket
-    from raiden.network.transport import MatrixTransport, UDPTransport
     from raiden.network.utils import get_free_port
-    from raiden.raiden_event_handler import RaidenEventHandler
-    from raiden.settings import (
-        DEFAULT_MATRIX_KNOWN_SERVERS,
-        DEFAULT_NAT_KEEPALIVE_RETRIES,
-        DEFAULT_SHUTDOWN_TIMEOUT,
-        ETHERSCAN_API,
-        INITIAL_PORT,
-        ORACLE_BLOCKNUMBER_DRIFT_TOLERANCE,
-    )
-    from raiden.storage.sqlite import RAIDEN_DB_VERSION, assert_sqlite_version
-    from raiden.tasks import check_gas_reserve, check_version
+    from raiden.settings import INITIAL_PORT
     from raiden.utils import (
         get_system_spec,
-        is_supported_client,
-        merge_dict,
-        networkid_is_known,
-        pex,
         split_endpoint,
-        typing,
     )
     from raiden.utils.cli import (
         ADDRESS_TYPE,
@@ -95,214 +40,20 @@ if True:
         NetworkChoiceType,
         PathRelativePath,
         apply_config_file,
-        get_matrix_servers,
         group,
         option,
         option_group,
     )
-    from raiden.utils.echo_node import EchoNode
     from raiden.utils.http import HTTPExecutor
-    from raiden.utils.runnable import Runnable
     from raiden_contracts.constants import (
-        ID_TO_NETWORKNAME,
-        ID_TO_NETWORK_CONFIG,
-        ChainId,
-        NetworkType,
-        START_QUERY_BLOCK_KEY,
         CONTRACT_ENDPOINT_REGISTRY,
-        CONTRACT_SECRET_REGISTRY,
         CONTRACT_TOKEN_NETWORK_REGISTRY,
     )
 
+    from .runners import NodeRunner, EchoNodeRunner
+    from .app import run_app
+
 log = structlog.get_logger(__name__)
-
-
-ETHEREUM_NODE_COMMUNICATION_ERROR = (
-    '\n'
-    'Could not contact the ethereum node through JSON-RPC.\n'
-    'Please make sure that JSON-RPC is enabled for these interfaces:\n'
-    '\n'
-    '    eth_*, net_*, web3_*\n'
-    '\n'
-    'geth: https://github.com/ethereum/go-ethereum/wiki/Management-APIs\n'
-)
-
-
-def check_synced(blockchain_service: BlockChainService, network_id_is_known: bool) -> None:
-    net_id = blockchain_service.network_id
-    if not network_id_is_known:
-        click.secho(
-            f'Your ethereum client is connected to a non-recognized private \n'
-            f'network with network-ID {net_id}. Since we can not check if the client \n'
-            f'is synced please restart raiden with the --no-sync-check argument.'
-            f'\n',
-            fg='red',
-        )
-        sys.exit(1)
-
-    try:
-        network = ID_TO_NETWORKNAME[ChainId(net_id)]
-    except (EthNodeCommunicationError, RequestException):
-        click.secho(
-            'Could not determine the network the ethereum node is connected.\n'
-            'Because of this there is no way to determine the latest\n'
-            'block with an oracle, and the events from the ethereum\n'
-            'node cannot be trusted. Giving up.\n',
-            fg='red',
-        )
-        sys.exit(1)
-
-    url = ETHERSCAN_API.format(
-        network=network if net_id != 1 else 'api',
-        action='eth_blockNumber',
-    )
-    wait_for_sync(
-        blockchain_service,
-        url=url,
-        tolerance=ORACLE_BLOCKNUMBER_DRIFT_TOLERANCE,
-        sleep=3,
-    )
-
-
-def check_discovery_registration_gas(
-        blockchain_service: BlockChainService,
-        account_address: typing.Address,
-) -> None:
-    discovery_tx_cost = blockchain_service.client.gas_price() * DISCOVERY_TX_GAS_LIMIT
-    account_balance = blockchain_service.client.balance(account_address)
-
-    # pylint: disable=no-member
-    if discovery_tx_cost > account_balance:
-        click.secho(
-            'Account has insufficient funds for discovery registration.\n'
-            'Needed: {} ETH\n'
-            'Available: {} ETH.\n'
-            'Please deposit additional funds into this account.'
-            .format(discovery_tx_cost / denoms.ether, account_balance / denoms.ether),
-            fg='red',
-        )
-        sys.exit(1)
-
-
-def etherscan_query_with_retries(
-        url: str,
-        sleep: float,
-        retries: int = 3,
-) -> int:
-    for _ in range(retries - 1):
-        try:
-            etherscan_block = to_int(hexstr=requests.get(url).json()['result'])
-        except (RequestException, ValueError, KeyError):
-            gevent.sleep(sleep)
-        else:
-            return etherscan_block
-
-    etherscan_block = to_int(hexstr=requests.get(url).json()['result'])
-    return etherscan_block
-
-
-def wait_for_sync_etherscan(
-        blockchain_service: BlockChainService,
-        url: str,
-        tolerance: int,
-        sleep: float,
-) -> None:
-    local_block = blockchain_service.client.block_number()
-    etherscan_block = etherscan_query_with_retries(url, sleep)
-    syncing_str = '\rSyncing ... Current: {} / Target: ~{}'
-
-    if local_block >= etherscan_block - tolerance:
-        return
-
-    print('Waiting for the ethereum node to synchronize. [Use ^C to exit]')
-    print(syncing_str.format(local_block, etherscan_block), end='')
-
-    for i in count():
-        sys.stdout.flush()
-        gevent.sleep(sleep)
-        local_block = blockchain_service.client.block_number()
-
-        # update the oracle block number sparsely to not spam the server
-        if local_block >= etherscan_block or i % 50 == 0:
-            etherscan_block = etherscan_query_with_retries(url, sleep)
-
-            if local_block >= etherscan_block - tolerance:
-                return
-
-        print(syncing_str.format(local_block, etherscan_block), end='')
-
-    # add a newline so that the next print will start have it's own line
-    print('')
-
-
-def wait_for_sync_rpc_api(
-        blockchain_service: BlockChainService,
-        sleep: float,
-) -> None:
-    if blockchain_service.is_synced():
-        return
-
-    print('Waiting for the ethereum node to synchronize [Use ^C to exit].')
-
-    for i in count():
-        if i % 3 == 0:
-            print('\r', end='')
-
-        print('.', end='')
-        sys.stdout.flush()
-
-        gevent.sleep(sleep)
-
-        if blockchain_service.is_synced():
-            return
-
-    # add a newline so that the next print will start have it's own line
-    print('')
-
-
-def wait_for_sync(
-        blockchain_service: BlockChainService,
-        url: str,
-        tolerance: int,
-        sleep: float,
-) -> None:
-    # print something since the actual test may take a few moments for the first
-    # iteration
-    print('Checking if the ethereum node is synchronized')
-
-    try:
-        wait_for_sync_etherscan(blockchain_service, url, tolerance, sleep)
-    except (RequestException, ValueError, KeyError):
-        print('Cannot use {}. Request failed'.format(url))
-        print('Falling back to eth_sync api.')
-
-        wait_for_sync_rpc_api(blockchain_service, sleep)
-
-
-def handle_contract_version_mismatch(name: str, address: typing.Address) -> None:
-    hex_addr = to_checksum_address(address)
-    click.secho(
-        f'Error: Provided {name} {hex_addr} contract version mismatch. '
-        'Please update your Raiden installation.',
-        fg='red',
-    )
-    sys.exit(1)
-
-
-def handle_contract_no_code(name: str, address: typing.Address) -> None:
-    hex_addr = to_checksum_address(address)
-    click.secho(f'Error: Provided {name} {hex_addr} contract does not contain code', fg='red')
-    sys.exit(1)
-
-
-def handle_contract_wrong_address(name: str, address: typing.Address) -> None:
-    hex_addr = to_checksum_address(address)
-    click.secho(
-        f'Error: Provided address {hex_addr} for {name} contract'
-        ' does not contain expected code.',
-        fg='red',
-    )
-    sys.exit(1)
 
 
 def options(func):
@@ -423,6 +174,12 @@ def options(func):
             '--accept-disclaimer',
             help='Bypass the experimental software disclaimer prompt',
             is_flag=True,
+        ),
+        option(
+            '--showconfig',
+            help='Shows all configuration values use by Raiden from various sources',
+            is_flag=True,
+            default=False,
         ),
         option_group(
             'Ethereum Node Options',
@@ -588,623 +345,6 @@ def options(func):
     for option_ in reversed(options_):
         func = option_(func)
     return func
-
-
-def run_app(
-        address,
-        keystore_path,
-        gas_price,
-        eth_rpc_endpoint,
-        registry_contract_address,
-        secret_registry_contract_address,
-        discovery_contract_address,
-        listen_address,
-        mapped_socket,
-        max_unresponsive_time,
-        api_address,
-        rpc,
-        sync_check,
-        console,
-        password_file,
-        web_ui,
-        datadir,
-        transport,
-        matrix_server,
-        network_id,
-        network_type,
-        extra_config=None,
-        **kwargs,
-):
-    # pylint: disable=too-many-locals,too-many-branches,too-many-statements,unused-argument
-
-    from raiden.app import App
-
-    if not assert_sqlite_version():
-        log.error('SQLite3 should be at least version {}'.format(
-            '{}.{}.{}'.format(*SQLITE_MIN_REQUIRED_VERSION),
-        ))
-        sys.exit(1)
-
-    if transport == 'udp' and not mapped_socket:
-        raise RuntimeError('Missing socket')
-
-    address_hex = to_normalized_address(address) if address else None
-    address_hex, privatekey_bin = prompt_account(address_hex, keystore_path, password_file)
-    address = to_canonical_address(address_hex)
-
-    (listen_host, listen_port) = split_endpoint(listen_address)
-    (api_host, api_port) = split_endpoint(api_address)
-
-    if datadir is None:
-        datadir = os.path.join(os.path.expanduser('~'), '.raiden')
-
-    config = deepcopy(App.DEFAULT_CONFIG)
-    if extra_config:
-        merge_dict(config, extra_config)
-
-    config['transport']['udp']['host'] = listen_host
-    config['transport']['udp']['port'] = listen_port
-    config['console'] = console
-    config['rpc'] = rpc
-    config['web_ui'] = rpc and web_ui
-    config['api_host'] = api_host
-    config['api_port'] = api_port
-    if mapped_socket:
-        config['socket'] = mapped_socket.socket
-        config['transport']['udp']['external_ip'] = mapped_socket.external_ip
-        config['transport']['udp']['external_port'] = mapped_socket.external_port
-    config['transport_type'] = transport
-    config['transport']['matrix']['server'] = matrix_server
-    config['transport']['udp']['nat_keepalive_retries'] = DEFAULT_NAT_KEEPALIVE_RETRIES
-    timeout = max_unresponsive_time / DEFAULT_NAT_KEEPALIVE_RETRIES
-    config['transport']['udp']['nat_keepalive_timeout'] = timeout
-
-    privatekey_hex = hexlify(privatekey_bin)
-    config['privatekey_hex'] = privatekey_hex
-
-    parsed_eth_rpc_endpoint = urlparse(eth_rpc_endpoint)
-    if not parsed_eth_rpc_endpoint.scheme:
-        eth_rpc_endpoint = f'http://{eth_rpc_endpoint}'
-
-    web3 = Web3(HTTPProvider(eth_rpc_endpoint))
-
-    try:
-        node_version = web3.version.node  # pylint: disable=no-member
-    except ConnectTimeout:
-        raise EthNodeCommunicationError("Couldn't connect to the ethereum node")
-
-    supported, _ = is_supported_client(node_version)
-    if not supported:
-        click.secho(
-            'You need a Byzantium enabled ethereum node. Parity >= 1.7.6 or Geth >= 1.7.2',
-            fg='red',
-        )
-        sys.exit(1)
-
-    rpc_client = JSONRPCClient(
-        web3,
-        privatekey_bin,
-        gas_price_strategy=gas_price,
-    )
-
-    blockchain_service = BlockChainService(privatekey_bin, rpc_client)
-
-    given_numeric_network_id = network_id.value if isinstance(network_id, ChainId) else network_id
-    node_numeric_network_id = blockchain_service.network_id
-    known_given_network_id = networkid_is_known(given_numeric_network_id)
-    known_node_network_id = networkid_is_known(node_numeric_network_id)
-    if known_given_network_id:
-        given_network_id = ChainId(given_numeric_network_id)
-    if known_node_network_id:
-        node_network_id = ChainId(node_numeric_network_id)
-
-    if node_numeric_network_id != given_numeric_network_id:
-        if known_given_network_id and known_node_network_id:
-            click.secho(
-                f"The chosen ethereum network '{given_network_id.name.lower()}' "
-                f"differs from the ethereum client '{node_network_id.name.lower()}'. "
-                "Please update your settings.",
-                fg='red',
-            )
-        else:
-            click.secho(
-                f"The chosen ethereum network id '{given_numeric_network_id}' differs "
-                f"from the ethereum client '{node_numeric_network_id}'. "
-                "Please update your settings.",
-                fg='red',
-            )
-        sys.exit(1)
-
-    config['chain_id'] = given_numeric_network_id
-
-    log.debug('Network type', type=network_type)
-    if network_type == 'main':
-        config['network_type'] = NetworkType.MAIN
-        # Forcing private rooms to true for the mainnet
-        config['transport']['matrix']['private_rooms'] = True
-    else:
-        config['network_type'] = NetworkType.TEST
-
-    network_type = config['network_type']
-    chain_config = {}
-    contract_addresses_known = False
-    contract_addresses = dict()
-    if node_network_id in ID_TO_NETWORK_CONFIG:
-        network_config = ID_TO_NETWORK_CONFIG[node_network_id]
-        not_allowed = (
-            NetworkType.TEST not in network_config and
-            network_type == NetworkType.TEST
-        )
-        if not_allowed:
-            click.secho(
-                'The chosen network {} has no test configuration but a test network type '
-                'was given. This is not allowed.'.format(
-                    ID_TO_NETWORKNAME[node_network_id],
-                ),
-                fg='red',
-            )
-            sys.exit(1)
-
-        if network_type in network_config:
-            chain_config = network_config[network_type]
-            contract_addresses = chain_config['contract_addresses']
-            contract_addresses_known = True
-
-    if sync_check:
-        check_synced(blockchain_service, known_node_network_id)
-
-    contract_addresses_given = (
-        registry_contract_address is not None and
-        secret_registry_contract_address is not None and
-        discovery_contract_address is not None
-    )
-
-    if not contract_addresses_given and not contract_addresses_known:
-        click.secho(
-            f"There are no known contract addresses for network id '{given_numeric_network_id}'. "
-            "Please provide them on the command line or in the configuration file.",
-            fg='red',
-        )
-        sys.exit(1)
-
-    try:
-        token_network_registry = blockchain_service.token_network_registry(
-            registry_contract_address or contract_addresses[CONTRACT_TOKEN_NETWORK_REGISTRY],
-        )
-    except ContractVersionMismatch:
-        handle_contract_version_mismatch('token network registry', registry_contract_address)
-    except AddressWithoutCode:
-        handle_contract_no_code('token network registry', registry_contract_address)
-    except AddressWrongContract:
-        handle_contract_wrong_address('token network registry', registry_contract_address)
-
-    try:
-        secret_registry = blockchain_service.secret_registry(
-            secret_registry_contract_address or contract_addresses[CONTRACT_SECRET_REGISTRY],
-        )
-    except ContractVersionMismatch:
-        handle_contract_version_mismatch('secret registry', secret_registry_contract_address)
-    except AddressWithoutCode:
-        handle_contract_no_code('secret registry', secret_registry_contract_address)
-    except AddressWrongContract:
-        handle_contract_wrong_address('secret registry', secret_registry_contract_address)
-
-    database_path = os.path.join(
-        datadir,
-        f'node_{pex(address)}',
-        f'netid_{given_numeric_network_id}',
-        f'network_{pex(token_network_registry.address)}',
-        f'v{RAIDEN_DB_VERSION}_log.db',
-    )
-    config['database_path'] = database_path
-
-    print(
-        '\nYou are connected to the \'{}\' network and the DB path is: {}'.format(
-            ID_TO_NETWORKNAME.get(given_network_id, given_numeric_network_id),
-            database_path,
-        ),
-    )
-
-    discovery = None
-    if transport == 'udp':
-        check_discovery_registration_gas(blockchain_service, address)
-        try:
-            dicovery_proxy = blockchain_service.discovery(
-                discovery_contract_address or contract_addresses[CONTRACT_ENDPOINT_REGISTRY],
-            )
-            discovery = ContractDiscovery(
-                blockchain_service.node_address,
-                dicovery_proxy,
-            )
-        except ContractVersionMismatch:
-            handle_contract_version_mismatch('discovery', discovery_contract_address)
-        except AddressWithoutCode:
-            handle_contract_no_code('discovery', discovery_contract_address)
-        except AddressWrongContract:
-            handle_contract_wrong_address('discovery', discovery_contract_address)
-
-        throttle_policy = TokenBucket(
-            config['transport']['udp']['throttle_capacity'],
-            config['transport']['udp']['throttle_fill_rate'],
-        )
-
-        transport = UDPTransport(
-            discovery,
-            mapped_socket.socket,
-            throttle_policy,
-            config['transport']['udp'],
-        )
-    elif transport == 'matrix':
-        if config['transport']['matrix'].get('available_servers') is None:
-            # fetch list of known servers from raiden-network/raiden-tranport repo
-            available_servers_url = DEFAULT_MATRIX_KNOWN_SERVERS[config['network_type']]
-            available_servers = get_matrix_servers(available_servers_url)
-            config['transport']['matrix']['available_servers'] = available_servers
-
-        try:
-            transport = MatrixTransport(config['transport']['matrix'])
-        except RaidenError as ex:
-            click.secho(f'FATAL: {ex}', fg='red')
-            sys.exit(1)
-    else:
-        raise RuntimeError(f'Unknown transport type "{transport}" given')
-
-    raiden_event_handler = RaidenEventHandler()
-
-    try:
-        start_block = chain_config.get(START_QUERY_BLOCK_KEY, 0)
-        raiden_app = App(
-            config=config,
-            chain=blockchain_service,
-            query_start_block=start_block,
-            default_registry=token_network_registry,
-            default_secret_registry=secret_registry,
-            transport=transport,
-            raiden_event_handler=raiden_event_handler,
-            discovery=discovery,
-        )
-    except RaidenError as e:
-        click.secho(f'FATAL: {e}', fg='red')
-        sys.exit(1)
-
-    try:
-        raiden_app.start()
-    except RuntimeError as e:
-        click.secho(f'FATAL: {e}', fg='red')
-        sys.exit(1)
-    except filelock.Timeout:
-        name_or_id = ID_TO_NETWORKNAME.get(given_network_id, given_numeric_network_id)
-        click.secho(
-            f'FATAL: Another Raiden instance already running for account {address_hex} on '
-            f'network id {name_or_id}',
-            fg='red',
-        )
-        sys.exit(1)
-
-    return raiden_app
-
-
-def prompt_account(address_hex, keystore_path, password_file):
-    accmgr = AccountManager(keystore_path)
-    if not accmgr.accounts:
-        click.secho(
-            'No Ethereum accounts found in the provided keystore directory {}. '
-            'Please provide a directory containing valid ethereum account '
-            'files.'.format(keystore_path),
-            fg='red',
-        )
-        sys.exit(1)
-
-    if not accmgr.address_in_keystore(address_hex):
-        # check if an address has been passed
-        if address_hex is not None:
-            click.secho(
-                f"Account '{address_hex}' could not be found on the system. Aborting ...",
-                fg='red',
-            )
-            sys.exit(1)
-
-        addresses = list(accmgr.accounts.keys())
-        formatted_addresses = [
-            '[{:3d}] - {}'.format(idx, to_checksum_address(addr))
-            for idx, addr in enumerate(addresses)
-        ]
-
-        should_prompt = True
-
-        print('The following accounts were found in your machine:')
-        print('')
-        print('\n'.join(formatted_addresses))
-        print('')
-
-        while should_prompt:
-            idx = click.prompt('Select one of them by index to continue', type=int)
-
-            if 0 <= idx < len(addresses):
-                should_prompt = False
-            else:
-                print('\nError: Provided index "{}" is out of bounds\n'.format(idx))
-
-        address_hex = addresses[idx]
-
-    password = None
-    if password_file:
-        password = password_file.read()
-        if password != '':
-            password = password.splitlines()[0]
-    if password is not None:
-        try:
-            privatekey_bin = accmgr.get_privkey(address_hex, password)
-        except ValueError:
-            # ValueError exception raised if the password is incorrect
-            click.secho(
-                f'Incorrect password for {address_hex} in file. Aborting ...',
-                fg='red',
-            )
-            sys.exit(1)
-    else:
-        unlock_tries = 3
-        while True:
-            try:
-                privatekey_bin = accmgr.get_privkey(address_hex)
-                break
-            except ValueError:
-                # ValueError exception raised if the password is incorrect
-                if unlock_tries == 0:
-                    click.secho(
-                        f'Exhausted passphrase unlock attempts for {address_hex}. Aborting ...',
-                        fg='red',
-                    )
-                    sys.exit(1)
-
-                print(
-                    'Incorrect passphrase to unlock the private key. {} tries remaining. '
-                    'Please try again or kill the process to quit. '
-                    'Usually Ctrl-c.'.format(unlock_tries),
-                )
-                unlock_tries -= 1
-
-    return address_hex, privatekey_bin
-
-
-class NodeRunner:
-    def __init__(self, options: Dict[str, Any], ctx):
-        self._options = options
-        self._ctx = ctx
-        self._raiden_api = None
-
-    @property
-    def _welcome_string(self):
-        return f"Welcome to Raiden, version {get_system_spec()['raiden']}!"
-
-    def _startup_hook(self):
-        """ Hook that is called after startup is finished. Intended for subclass usage. """
-        pass
-
-    def _shutdown_hook(self):
-        """ Hook that is called just before shutdown. Intended for subclass usage. """
-        pass
-
-    def run(self):
-        click.secho(self._welcome_string, fg='green')
-        click.secho(
-            textwrap.dedent(
-                '''\
-                ----------------------------------------------------------------------
-                | This is an Alpha version of experimental open source software      |
-                | released under the MIT license and may contain errors and/or bugs. |
-                | Use of the software is at your own risk and discretion. No         |
-                | guarantee whatsoever is made regarding its suitability for your    |
-                | intended purposes and its compliance with applicable law and       |
-                | regulations. It is up to the user to determine the software´s      |
-                | quality and suitability and whether its use is compliant with its  |
-                | respective regulatory regime.                                      |
-                |                                                                    |
-                | Privacy notice: Please be aware, that by using the Raiden Client,  |
-                | your Ethereum address, channels, channel deposits, settlements and |
-                | the Ethereum address of your settlement counterparty will be       |
-                | stored on the Ethereum chain, i.e. on servers of Ethereum node     |
-                | operators and ergo made publicly available. The same will also be  |
-                | stored on systems of parties running other Raiden nodes connected  |
-                | to the same token network.                                         |
-                |                                                                    |
-                | Also be aware, that data on individual Raiden token transfers will |
-                | be made available via the Matrix protocol to the recipient,        |
-                | intermediating nodes of a specific transfer as well as to the      |
-                | Matrix server operators.                                           |
-                ----------------------------------------------------------------------''',
-            ),
-            fg='yellow',
-        )
-        if not self._options['accept_disclaimer']:
-            click.confirm('\nHave you read and acknowledged the above disclaimer?', abort=True)
-
-        configure_logging(
-            self._options['log_config'],
-            log_json=self._options['log_json'],
-            log_file=self._options['log_file'],
-            disable_debug_logfile=self._options['disable_debug_logfile'],
-        )
-
-        if self._options['config_file']:
-            log.debug('Using config file', config_file=self._options['config_file'])
-
-        # TODO:
-        # - Ask for confirmation to quit if there are any locked transfers that did
-        # not timeout.
-        try:
-            if self._options['transport'] == 'udp':
-                (listen_host, listen_port) = split_endpoint(self._options['listen_address'])
-                try:
-                    with SocketFactory(
-                        listen_host, listen_port, strategy=self._options['nat'],
-                    ) as mapped_socket:
-                        self._options['mapped_socket'] = mapped_socket
-                        app = self._run_app()
-
-                except RaidenServicePortInUseError:
-                    click.secho(
-                        'ERROR: Address %s:%s is in use. '
-                        'Use --listen-address <host:port> to specify port to listen on.' %
-                        (listen_host, listen_port),
-                        fg='red',
-                    )
-                    sys.exit(1)
-            elif self._options['transport'] == 'matrix':
-                self._options['mapped_socket'] = None
-                app = self._run_app()
-            else:
-                # Shouldn't happen
-                raise RuntimeError(f"Invalid transport type '{self._options['transport']}'")
-            app.stop()
-        except (ReplacementTransactionUnderpriced, TransactionAlreadyPending) as e:
-            click.secho(
-                '{}. Please make sure that this Raiden node is the '
-                'only user of the selected account'.format(str(e)),
-                fg='red',
-            )
-            sys.exit(1)
-
-    def _run_app(self):
-        from raiden.ui.console import Console
-        from raiden.api.python import RaidenAPI
-
-        # this catches exceptions raised when waiting for the stalecheck to complete
-        try:
-            app_ = run_app(**self._options)
-        except (EthNodeCommunicationError, RequestsConnectionError):
-            print(ETHEREUM_NODE_COMMUNICATION_ERROR)
-            sys.exit(1)
-
-        tasks = [app_.raiden]  # RaidenService takes care of Transport and AlarmTask
-
-        domain_list = []
-        if self._options['rpccorsdomain']:
-            if ',' in self._options['rpccorsdomain']:
-                for domain in self._options['rpccorsdomain'].split(','):
-                    domain_list.append(str(domain))
-            else:
-                domain_list.append(str(self._options['rpccorsdomain']))
-
-        self._raiden_api = RaidenAPI(app_.raiden)
-
-        if self._options['rpc']:
-            rest_api = RestAPI(self._raiden_api)
-            api_server = APIServer(
-                rest_api,
-                cors_domain_list=domain_list,
-                web_ui=self._options['web_ui'],
-                eth_rpc_endpoint=self._options['eth_rpc_endpoint'],
-            )
-            (api_host, api_port) = split_endpoint(self._options['api_address'])
-
-            try:
-                api_server.start(api_host, api_port)
-            except APIServerPortInUseError:
-                click.secho(
-                    f'ERROR: API Address {api_host}:{api_port} is in use. '
-                    f'Use --api-address <host:port> to specify a different port.',
-                    fg='red',
-                )
-                sys.exit(1)
-
-            print(
-                'The Raiden API RPC server is now running at http://{}:{}/.\n\n'
-                'See the Raiden documentation for all available endpoints at\n'
-                'http://raiden-network.readthedocs.io/en/stable/rest_api.html'.format(
-                    api_host,
-                    api_port,
-                ),
-            )
-            tasks.append(api_server)
-
-        if self._options['console']:
-            console = Console(app_)
-            console.start()
-            tasks.append(console)
-
-        # spawn a greenlet to handle the version checking
-        version = get_system_spec()['raiden']
-        if version is not None:
-            tasks.append(gevent.spawn(check_version, version))
-
-        # spawn a greenlet to handle the gas reserve check
-        tasks.append(gevent.spawn(check_gas_reserve, app_.raiden))
-
-        self._startup_hook()
-
-        # wait for interrupt
-        event = AsyncResult()
-
-        def sig_set(sig=None, _frame=None):
-            event.set(sig)
-
-        gevent.signal(signal.SIGQUIT, sig_set)
-        gevent.signal(signal.SIGTERM, sig_set)
-        gevent.signal(signal.SIGINT, sig_set)
-
-        # quit if any task exits, successfully or not
-        for task in tasks:
-            task.link(event)
-
-        try:
-            event.get()
-            print('Signal received. Shutting down ...')
-        except (EthNodeCommunicationError, RequestsConnectionError):
-            print(ETHEREUM_NODE_COMMUNICATION_ERROR)
-            sys.exit(1)
-        except RaidenError as ex:
-            click.secho(f'FATAL: {ex}', fg='red')
-        except Exception as ex:
-            with NamedTemporaryFile(
-                'w',
-                prefix=f'raiden-exception-{datetime.utcnow():%Y-%m-%dT%H-%M}',
-                suffix='.txt',
-                delete=False,
-            ) as traceback_file:
-                traceback.print_exc(file=traceback_file)
-                click.secho(
-                    f'FATAL: An unexpected exception occured. '
-                    f'A traceback has been written to {traceback_file.name}\n'
-                    f'{ex}',
-                    fg='red',
-                )
-        finally:
-            self._shutdown_hook()
-
-            def stop_task(task):
-                try:
-                    if isinstance(task, Runnable):
-                        task.stop()
-                    else:
-                        task.kill()
-                finally:
-                    task.get()  # re-raise
-
-            gevent.joinall(
-                [gevent.spawn(stop_task, task) for task in tasks],
-                app_.config.get('shutdown_timeout', DEFAULT_SHUTDOWN_TIMEOUT),
-                raise_error=True,
-            )
-
-        return app_
-
-
-class EchoNodeRunner(NodeRunner):
-    def __init__(self, options: Dict[str, Any], ctx, token_address: typing.TokenAddress):
-        super().__init__(options, ctx)
-        self._token_address = token_address
-        self._echo_node = None
-
-    @property
-    def _welcome_string(self):
-        return '{} [ECHO NODE]'.format(super(EchoNodeRunner, self)._welcome_string)
-
-    def _startup_hook(self):
-        self._echo_node = EchoNode(self._raiden_api, self._token_address)
-
-    def _shutdown_hook(self):
-        self._echo_node.stop()
 
 
 @group(invoke_without_command=True, context_settings={'max_content_width': 120})
